improve documents

+++ build.sh
@@ -8,6 +8,8 @@ cd ${DIR}
 
 rm -rf _book
 
+npm install markdown-parser2
+
 for d in */; do
 
   if [ ! -f ${d}/book.json ]; then
@@ -15,14 +17,16 @@ for d in */; do
   fi
 
   echo "Generate PDF for book ${d}"
-
-  gitbook install --log error ${d} >/dev/null
+  cp ./gen_summary.js ${d}
+  cd ${d} && node gen_summary.js && rm gen_summary.js && cd ..
+  test ! -e node_modules && gitbook install --log error ${d}
   gitbook pdf --log warn ${d} "${d}$(basename ${d}).pdf"
 done
 
 echo "Generate website"
 
-gitbook install --log error
+test ! -e node_modules && gitbook install --log error
+node gen_summary.js
 gitbook build --log warn
 
 find _book -name "SUMMARY.md" | xargs rm -f
